Some unrelated changes that shouldn't have been in a single commit...
1) Reformat code in pyobjc-core: TABs in ObjC sources are replaced by SPACEs

2) Move NSDecimal wrappers to pyobjc-core both to reduce unwanted
   dependency inversion and because NSDecimal is a struct type that
   cannot be handled at all by libffi and hence requires special support.

3) Some small bugfixes in error paths.

+++ pyobjc-core/Modules/objc/OC_PythonDate.m
@@ -3,7 +3,7 @@
 
 static PyObject* datetime_types = NULL;
 
-@implementation OC_PythonDate 
+@implementation OC_PythonDate
 
 + (instancetype)depythonifyObject:(PyObject*)object
 {
@@ -85,8 +85,8 @@ static PyObject* datetime_types = NULL;
     return value;
 }
 
--(BOOL)supportsWeakPointers { 
-    return YES; 
+-(BOOL)supportsWeakPointers {
+    return YES;
 }
 
 -(oneway void)release
@@ -119,7 +119,7 @@ static PyObject* datetime_types = NULL;
 }
 
 
-/* 
+/*
  * Helper method for initWithCoder, needed to deal with
  * recursive objects (e.g. o.value = o)
  */
@@ -207,13 +207,13 @@ static PyObject* datetime_types = NULL;
                 }
                 offset = offset / 60; /* Seconds to minutes */
 
-                int minutes = offset % 60;
-                int hours = offset / 60;
+                NSInteger minutes = offset % 60;
+                NSInteger hours = offset / 60;
 
 
 
-                snprintf(buf, sizeof(buf), "%%Y-%%m-%%d %%H:%%M:%%S %c%02d%02d",
-                    posneg, hours, minutes);
+                snprintf(buf, sizeof(buf), "%%Y-%%m-%%d %%H:%%M:%%S %c%02ld%02ld",
+                    posneg, (long)hours, (long)minutes);
                 v = PyObject_CallMethod(value, "strftime", "s", buf);
                 if (v == NULL) {
                     /* Raise ObjC exception */
@@ -225,7 +225,7 @@ static PyObject* datetime_types = NULL;
             }
 
         PyObjC_END_WITH_GIL
-    } 
+    }
     return oc_value;
 }
